requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
backports.entry-points-selectable==1.1.1
billiard==3.5.0.5
cached-property==1.5.2
celery==4.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
cryptography==3.1
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.4
django-allauth==0.39.1
django-bootstrap-datetimepicker==1.5
django-bootstrap3==11.1.0
django-cors-headers==3.5.0
django-debug-toolbar==2.0
django-extensions==2.2.1
django-filter==2.2.0
django-floppyforms==1.9.0
django-geojson==2.11.0
django-jet==1.0.8
django-jquery==3.1.0
django-json==0.27
django-leaflet==0.24.0
django-map-widgets==0.3.0
django-reversion==4.0.1
django-simple-menu==1.2.1
django-smart-selects==1.5.9
django-tables2==2.1.0
django-webpack-loader==0.7.0
djangorestframework==3.12.1
filelock==3.0.12
#GDAL==2.4.4
graphviz==0.18
gunicorn==19.9.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
iniconfig==1.0.1
isort==5.5.4
kombu==4.6.11
lazy-object-proxy==1.4.3
marshmallow==2.15.6
mccabe==0.6.1
mock==3.0.5
more-itertools==8.5.0
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
platformdirs==2.4.0
pluggy==0.13.1
prompt-toolkit==3.0.22
psycopg2==2.8.3
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.0.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
reportlab==3.5.51
requests==2.24.0
requests-oauthlib==1.3.0
reversion==0.2
six==1.15.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
transbank-sdk==1.2.0
typed-ast==1.4.1
typing-extensions==3.10.0.2
urllib3==1.25.10
utils==0.9.0
vine==1.3.0
virtualenv==20.0.31
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0
